+++ src/pages/emails/EmailRecipientsModal.jsx
@@ -1,64 +1,11 @@
 import React from 'react';
-import { useNavigate } from 'react-router-dom';
-import styled from 'styled-components';
+// import styled from 'styled-components';
+import Modal from '../../components/Modal';
 
 export default function EmailRecipientsModal() {
-  const navigate = useNavigate();
-
-  const handleCloseButtonClick = () => {
-    navigate(-1);
-  };
-
   return (
-    <ModalBackground>
-      <ModalContainer>
-        <Title>수신자 목록</Title>
-        <Content />
-        <CloseButton onClick={handleCloseButtonClick}>닫기</CloseButton>
-      </ModalContainer>
-    </ModalBackground>
+    <Modal title="수신자 목록">
+      <div>하이하이</div>
+    </Modal>
   );
 }
-
-const ModalBackground = styled.div`
-  width: 100vw;
-  height: 100%;
-  position: fixed;
-  top: 0;
-  background-color: rgba(0, 0, 0, 0.5);
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-const ModalContainer = styled.div`
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  width: 500px;
-  height: 650px;
-  background-color: white;
-`;
-
-const Title = styled.span`
-  padding: 40px 0;
-  font-size: 1.125rem;
-  font-weight: 500;
-`;
-
-const Content = styled.div`
-  margin-bottom: 30px;
-  width: 426px;
-  height: 450px;
-  border: 1px solid black;
-`;
-
-const CloseButton = styled.button`
-  width: 80px;
-  height: 40px;
-  border-radius: 5px;
-  background-color: #ffdf2b;
-  font-size: 0.875rem;
-  font-weight: 500;
-  text-align: center;
-`;
